telegram-bot/requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
click==8.1.3
loguru==0.7.0
python-dotenv==1.0.0
python-dateutil~=2.8.2